+++ Frends.Redis.GetValue/Frends.Redis.GetValue/Definitions/ConnectionMethod.cs
@@ -1,0 +1,7 @@
+﻿namespace Frends.Redis.GetValue.Definitions;
+
+public enum ConnectionMethod
+{
+    SimpleConnectionString = 1,
+    MicrosoftEntraId = 2,
+}